+++ src/deps.jl
@@ -134,11 +134,11 @@ function current_packages()
     b = backend()
     if b in CONDA_BACKENDS
         cmd = conda_cmd(`list -p $(envdir()) --json`)
-        pkglist = JSON3.read(cmd)
+        pkglist = JSON.parse(cmd)
     elseif b in PIXI_BACKENDS
         cmd =
             pixi_cmd(`list --manifest-path $(joinpath(STATE.meta_dir, "pixi.toml")) --json`)
-        pkglist = JSON3.read(cmd)
+        pkglist = JSON.parse(cmd)
         pkglist = [pkg for pkg in pkglist if pkg.kind == "conda"]
     end
     Dict(normalise_pkg(pkg.name) => pkg for pkg in pkglist)
@@ -149,12 +149,12 @@ function current_pip_packages()
     if b in CONDA_BACKENDS
         pkglist = withenv() do
             cmd = `$(which("pip")) list --format=json`
-            JSON3.read(cmd)
+            JSON.parse(cmd)
         end
     elseif b in PIXI_BACKENDS
         cmd =
             pixi_cmd(`list --manifest-path $(joinpath(STATE.meta_dir, "pixi.toml")) --json`)
-        pkglist = JSON3.read(cmd)
+        pkglist = JSON.parse(cmd)
         pkglist = [pkg for pkg in pkglist if pkg.kind == "pypi"]
     end
     Dict(normalise_pip_pkg(pkg.name) => pkg for pkg in pkglist)