+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 * If the `env` preference (or `JULIA_CONDAPKG_ENV`) can now be a relative path, which is
   taken to be relative to the current Julia project.
+* Internal: move from JSON3 to JSON v1.
 
 ## 0.2.33 (2025-09-18)
 * Add `[dev]` section to CondaPkg.toml, specifying development dependencies which are only